requirements.txt
apache-airflow[s3,docker]==1.10.10
Flask==1.1.1
psycopg2==2.8.4
dj-database-url==0.5.0
Flask-Bcrypt==0.7.1
